+++ Parser/ast.h
@@ -153,7 +153,7 @@ extern AST *ast_ident(token t, std::string str);
 
 // Return a (pointer to a) fresh AST for an (signed) number expression
 // with the given value
-extern AST *ast_number(token t, short int value);
+extern AST *ast_number(token t, float value);
 
 // Return an AST list that is empty
 extern AST_list ast_list_empty_list();
